+++ apps/next-app/CHANGELOG.md
@@ -1,5 +1,47 @@
 # next-app
 
+## 0.1.21
+
+### Patch Changes
+
+- Updated dependencies [84894ec]
+  - @nation-a/ui@0.11.0
+
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [6665698]
+  - @nation-a/ui@0.10.8
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [8976f4f]
+  - @nation-a/ui@0.10.7
+
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [dee01d5]
+  - @nation-a/ui@0.10.6
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [8a76f41]
+  - @nation-a/ui@0.10.5
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [a32a8cf]
+  - @nation-a/ui@0.10.4
+
 ## 0.1.15
 
 ### Patch Changes