+++ apps/test-app/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @nation-a/test-app
 
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies [84894ec]
+  - @nation-a/ui@0.11.0
+
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [6665698]
+  - @nation-a/ui@0.10.8
+
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [8976f4f]
+  - @nation-a/ui@0.10.7
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [dee01d5]
+  - @nation-a/ui@0.10.6
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [8a76f41]
+  - @nation-a/ui@0.10.5
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [a32a8cf]
+  - @nation-a/ui@0.10.4
+
 ## 0.0.24
 
 ### Patch Changes